Move register settigns calls out of admin
