+++ includes/admin/functions-settings.php
@@ -8,29 +8,8 @@
 namespace ucare;
 
 
-// Register settings
-add_action( 'init', 'ucare\register_settings' );
-
 // Add settings fields
 add_action( 'admin_init', 'ucare\add_settings_fields', 100 );
-
-
-/**
- * Action to register settings with the settings API.
- *
- * @action init
- *
- * @since 1.4.2
- * @return void
- */
-function register_settings() {
-
-    register_setting( 'uc-advanced', Options::TEMPLATE_PAGE_ID, array(
-        'type' => 'integer',
-        'sanitize_callback' => 'ucare\sanitize_post_id',
-    ) );
-
-}
 
 
 /**
